Issue 33: Move LongInt to delaunay/longint dir
