+++ include/delaunay/pointsetarray.h
@@ -3,8 +3,9 @@
 
 #include <vector>
 
-#include "delaunay/li.h"
-#include "delaunay/lmath.h"
+#include "delaunay/longint/li.h"
+#include "delaunay/longint/lmath.h"
+
 #include "delaunay/polygon.h"
