+++ src/delaunay/delaunay.cpp
@@ -6,8 +6,9 @@
 
 #include <iostream>
 
-#include "delaunay/li.h"
-#include "delaunay/lmath.h"
+#include "delaunay/longint/li.h"
+#include "delaunay/longint/lmath.h"
+
 #include "delaunay/pointsetarray.h"
 #include "delaunay/triangle.h"
 #include "delaunay/trianglegeometry.h"
