+++ src/delaunay/triangulation.cpp
@@ -4,7 +4,8 @@
 
 #include <iostream>
 
-#include "delaunay/li.h"
+#include "delaunay/longint/li.h"
+
 #include "delaunay/pointsetarray.h"
 
 using std::unique_ptr;
